+++ lib/danger/ci_source/bitrise.rb
@@ -1,5 +1,5 @@
 # http://devcenter.bitrise.io/docs/available-environment-variables
-require "danger/request_sources/github"
+require "danger/request_sources/github/github"
 require "danger/request_sources/gitlab"
 
 module Danger