+++ lib/danger/request_sources/github/github.rb
@@ -2,7 +2,8 @@
 require "octokit"
 require "danger/helpers/comments_helper"
 require "danger/helpers/comment"
-
+require "danger/request_sources/github/github_review"
+require "danger/request_sources/github/octokit_pr_review.rb"
 require "danger/request_sources/support/get_ignored_violation"
 
 module Danger
@@ -61,6 +62,16 @@ module Danger
 
       def pr_diff
         @pr_diff ||= client.pull_request(ci_source.repo_slug, ci_source.pull_request_id, accept: "application/vnd.github.v3.diff")
+      end
+
+      def review
+        return @review unless @review.nil?
+        @review = client.pull_request_reviews(ci_source.repo_slug, ci_source.pull_request_id)
+          .map { |review_json| Danger::RequestSources::GitHubSource::Review.new(client, ci_source, review_json) }
+          .select(&:generated_by_danger?)
+          .last
+        @review ||= Danger::RequestSources::GitHubSource::Review.new(client, ci_source)
+        @review
       end
 
       def setup_danger_branches